revert refactoring of EventSourceListener class
