+++ langchain4j-anthropic/src/main/java/dev/langchain4j/model/anthropic/internal/client/DefaultAnthropicClient.java
@@ -1,29 +1,46 @@
 package dev.langchain4j.model.anthropic.internal.client;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import dev.langchain4j.agent.tool.ToolExecutionRequest;
 import dev.langchain4j.data.message.AiMessage;
 import dev.langchain4j.internal.Utils;
 import dev.langchain4j.model.StreamingResponseHandler;
-import dev.langchain4j.model.anthropic.internal.api.AnthropicApi;
-import dev.langchain4j.model.anthropic.internal.api.AnthropicCreateMessageRequest;
-import dev.langchain4j.model.anthropic.internal.api.AnthropicCreateMessageResponse;
-import dev.langchain4j.model.anthropic.internal.api.AnthropicToolResultContent;
-import dev.langchain4j.model.anthropic.internal.api.AnthropicToolUseContent;
+import dev.langchain4j.model.anthropic.internal.api.*;
+import dev.langchain4j.model.output.Response;
+import dev.langchain4j.model.output.TokenUsage;
 import okhttp3.OkHttpClient;
 import okhttp3.ResponseBody;
+import okhttp3.sse.EventSource;
+import okhttp3.sse.EventSourceListener;
 import okhttp3.sse.EventSources;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import retrofit2.Call;
 import retrofit2.Retrofit;
 import retrofit2.converter.jackson.JacksonConverterFactory;
 
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicReference;
+import java.util.concurrent.locks.ReentrantLock;
+import java.util.stream.Collectors;
 
 import static com.fasterxml.jackson.databind.SerializationFeature.INDENT_OUTPUT;
+import static dev.langchain4j.internal.Utils.isNotNullOrEmpty;
 import static dev.langchain4j.internal.Utils.isNullOrBlank;
 import static dev.langchain4j.internal.Utils.isNullOrEmpty;
 import static dev.langchain4j.internal.ValidationUtils.ensureNotBlank;
+import static dev.langchain4j.model.anthropic.internal.mapper.AnthropicMapper.toFinishReason;
+import static java.util.Collections.synchronizedList;
 
 public class DefaultAnthropicClient extends AnthropicClient {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(DefaultAnthropicClient.class);
 
     private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper().enable(INDENT_OUTPUT);
 
@@ -115,7 +132,240 @@ public class DefaultAnthropicClient extends AnthropicClient {
 
     @Override
     public void createMessage(AnthropicCreateMessageRequest request, StreamingResponseHandler<AiMessage> handler) {
-        AnthropicEventSourceListener eventSourceListener = new AnthropicEventSourceListener(handler, logResponses);
+
+
+        EventSourceListener eventSourceListener = new EventSourceListener() {
+
+            private final ReentrantLock lock = new ReentrantLock();
+            final List<String> contents = synchronizedList(new ArrayList<>());
+            volatile StringBuffer currentContentBuilder = new StringBuffer();
+            private final AtomicReference<AnthropicContentBlockType> currentContentBlockStartType = new AtomicReference<>();
+
+            final AtomicInteger inputTokenCount = new AtomicInteger();
+            final AtomicInteger outputTokenCount = new AtomicInteger();
+            private final Map<Integer, AnthropicToolExecutionRequestBuilder> toolExecutionRequestBuilderMap = new ConcurrentHashMap<>();
+            AtomicReference<String> responseId = new AtomicReference<>();
+            AtomicReference<String> responseModel = new AtomicReference<>();
+
+            volatile String stopReason;
+
+            private StringBuffer currentContentBuilder() {
+                lock.lock();
+                try {
+                    return currentContentBuilder;
+                } finally {
+                    lock.unlock();
+                }
+            }
+
+            private void setCurrentContentBuilder(StringBuffer stringBuffer) {
+                lock.lock();
+                try {
+                    currentContentBuilder = stringBuffer;
+                } finally {
+                    lock.unlock();
+                }
+            }
+
+            @Override
+            public void onOpen(EventSource eventSource, okhttp3.Response response) {
+                if (logResponses) {
+                    LOGGER.debug("onOpen()");
+                }
+            }
+
+            @Override
+            public void onEvent(EventSource eventSource, String id, String type, String dataString) {
+                if (logResponses) {
+                    LOGGER.debug("onEvent() type: '{}', data: {}", type, dataString);
+                }
+
+                try {
+                    AnthropicStreamingData data = OBJECT_MAPPER.readValue(dataString, AnthropicStreamingData.class);
+
+                    if ("message_start".equals(type)) {
+                        handleMessageStart(data);
+                    } else if ("content_block_start".equals(type)) {
+                        handleContentBlockStart(data);
+                    } else if ("content_block_delta".equals(type)) {
+                        handleContentBlockDelta(data);
+                    } else if ("content_block_stop".equals(type)) {
+                        handleContentBlockStop();
+                    } else if ("message_delta".equals(type)) {
+                        handleMessageDelta(data);
+                    } else if ("message_stop".equals(type)) {
+                        handleMessageStop();
+                    } else if ("error".equals(type)) {
+                        handleError(dataString);
+                    }
+                } catch (Exception e) {
+                    handler.onError(e);
+                }
+            }
+
+            private void handleMessageStart(AnthropicStreamingData data) {
+                AnthropicResponseMessage message = data.message;
+                if (message != null) {
+                    if (message.usage != null) {
+                        handleUsage(message.usage);
+                    }
+                    if (message.id != null) {
+                        responseId.set(message.id);
+                    }
+                    if (message.model != null) {
+                        responseModel.set(message.model);
+                    }
+                }
+            }
+
+            private void handleUsage(AnthropicUsage usage) {
+                if (usage.inputTokens != null) {
+                    this.inputTokenCount.addAndGet(usage.inputTokens);
+                }
+                if (usage.outputTokens != null) {
+                    this.outputTokenCount.addAndGet(usage.outputTokens);
+                }
+            }
+
+            private void handleContentBlockStart(AnthropicStreamingData data) {
+                if (data.contentBlock == null) {
+                    return;
+                }
+
+                currentContentBlockStartType.set(data.contentBlock.type);
+
+                if (currentContentBlockStartType.get() == AnthropicContentBlockType.TEXT) {
+                    String text = data.contentBlock.text;
+                    if (isNotNullOrEmpty(text)) {
+                        currentContentBuilder().append(text);
+                        handler.onNext(text);
+                    }
+                } else if (currentContentBlockStartType.get() == AnthropicContentBlockType.TOOL_USE) {
+                    toolExecutionRequestBuilderMap.putIfAbsent(
+                            data.index,
+                            new AnthropicToolExecutionRequestBuilder(data.contentBlock.id, data.contentBlock.name)
+                    );
+                }
+            }
+
+            private void handleContentBlockDelta(AnthropicStreamingData data) {
+                if (data.delta == null) {
+                    return;
+                }
+
+                if (currentContentBlockStartType.get() == AnthropicContentBlockType.TEXT) {
+                    String text = data.delta.text;
+                    if (isNotNullOrEmpty(text)) {
+                        currentContentBuilder().append(text);
+                        handler.onNext(text);
+                    }
+                } else if (currentContentBlockStartType.get() == AnthropicContentBlockType.TOOL_USE) {
+                    String partialJson = data.delta.partialJson;
+                    if (isNotNullOrEmpty(partialJson)) {
+                        Integer toolExecutionsIndex = data.index;
+                        if (toolExecutionsIndex != null) {
+                            AnthropicToolExecutionRequestBuilder toolExecutionRequestBuilder = toolExecutionRequestBuilderMap.get(toolExecutionsIndex);
+                            toolExecutionRequestBuilder.appendArguments(partialJson);
+                        }
+                    }
+                }
+            }
+
+            private void handleContentBlockStop() {
+                contents.add(currentContentBuilder().toString());
+                setCurrentContentBuilder(new StringBuffer());
+            }
+
+            private void handleMessageDelta(AnthropicStreamingData data) {
+                if (data.delta != null) {
+                    AnthropicDelta delta = data.delta;
+                    if (delta.stopReason != null) {
+                        this.stopReason = delta.stopReason;
+                    }
+                }
+                if (data.usage != null) {
+                    handleUsage(data.usage);
+                }
+            }
+
+            private void handleMessageStop() {
+                Response<AiMessage> response = build();
+                handler.onComplete(response);
+            }
+
+            private Response<AiMessage> build() {
+                if (!toolExecutionRequestBuilderMap.isEmpty()) {
+                    List<ToolExecutionRequest> toolExecutionRequests = toolExecutionRequestBuilderMap
+                            .values().stream()
+                            .map(AnthropicToolExecutionRequestBuilder::build)
+                            .collect(Collectors.toList());
+                    return Response.from(
+                            AiMessage.from(toolExecutionRequests),
+                            new TokenUsage(inputTokenCount.get(), outputTokenCount.get()),
+                            toFinishReason(stopReason),
+                            createMetadata()
+                    );
+                }
+
+                String content = String.join("\n", contents);
+                if (!content.isEmpty()) {
+                    return Response.from(
+                            AiMessage.from(content),
+                            new TokenUsage(inputTokenCount.get(), outputTokenCount.get()),
+                            toFinishReason(stopReason),
+                            createMetadata()
+                    );
+                }
+
+                return null;
+            }
+
+            private Map<String, Object> createMetadata() {
+                Map<String, Object> metadata = new HashMap<>();
+                if (responseId.get() != null) {
+                    metadata.put("id", responseId.get());
+                }
+                if (responseModel.get() != null) {
+                    metadata.put("model", responseModel.get());
+                }
+                return metadata;
+            }
+
+            private void handleError(String dataString) {
+                handler.onError(new AnthropicHttpException(null, dataString));
+            }
+
+            @Override
+            public void onFailure(EventSource eventSource, Throwable t, okhttp3.Response response) {
+                if (logResponses) {
+                    LOGGER.debug("onFailure()", t);
+                }
+
+                if (t != null) {
+                    handler.onError(t);
+                }
+
+                if (response != null) {
+                    try (ResponseBody responseBody = response.body()) {
+                        if (responseBody != null) {
+                            handler.onError(new AnthropicHttpException(response.code(), responseBody.string()));
+                        } else {
+                            handler.onError(new AnthropicHttpException(response.code(), null));
+                        }
+                    } catch (IOException e) {
+                        handler.onError(new AnthropicHttpException(response.code(), "[error reading response body]"));
+                    }
+                }
+            }
+
+            @Override
+            public void onClosed(EventSource eventSource) {
+                if (logResponses) {
+                    LOGGER.debug("onClosed()");
+                }
+            }
+        };
+
         Call<ResponseBody> call = anthropicApi.streamMessage(apiKey, version, request);
         EventSources.createFactory(okHttpClient).newEventSource(call.request(), eventSourceListener);
     }
